+++ plugins/postcss-place/.tape.js
@@ -2,6 +2,12 @@ module.exports = {
 	'postcss-place': {
 		'basic': {
 			message: 'supports basic usage'
+		},
+		'basic:preserve-false': {
+			message: 'supports { preserve: false } usage',
+			options: {
+				preserve: false
+			}
 		}
 	}
 };