+++ plugins/postcss-place/CHANGELOG.md
@@ -1,4 +1,9 @@
-# Changes to Place
+# Changes to PostCSS Place Properties
+
+### 3.0.0 (May 8, 2018)
+
+- Updated: `postcss-selector-parser` to v4.0.0 (major)
+- Updated: `postcss` to v6.0.22 (patch)
 
 ### 2.0.0 (June 30, 2017)